refactor!: refactored module structure
Renamed mtgproxies to mtg_proxies and moved scryfall to mtg_proxies.scryfall
