+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+## 0.2.0 (2026-02-15)
+
+### Refactor
+
+- refactored module structure
