+++ convert.py
@@ -1,7 +1,7 @@
 import argparse
 from pathlib import Path
 
-from mtgproxies.cli import parse_decklist_spec
+from mtg_proxies.cli import parse_decklist_spec
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Convert a decklist from text format to arena format or vice-versa.")
