+++ mtg_proxies/__init__.py
@@ -1,0 +1,10 @@
+import mtg_proxies.scryfall as scryfall
+from mtg_proxies.print_cards import print_cards_fpdf, print_cards_matplotlib
+from mtg_proxies.scans import fetch_scans_scryfall
+
+__all__ = [
+    "fetch_scans_scryfall",
+    "print_cards_fpdf",
+    "print_cards_matplotlib",
+    "scryfall",
+]
