+++ mtg_proxies/decklists/archidekt/__init__.py
@@ -1,0 +1,5 @@
+from mtg_proxies.decklists.archidekt.archidekt import parse_decklist
+
+__all__ = [
+    "parse_decklist",
+]
