+++ mtg_proxies/decklists/manastack/__init__.py
@@ -1,0 +1,5 @@
+from mtg_proxies.decklists.manastack.manastack import parse_decklist
+
+__all__ = [
+    "parse_decklist",
+]
