+++ mtg_proxies/decklists/manastack/manastack.py
@@ -1,0 +1,58 @@
+from __future__ import annotations
+
+from collections.abc import Sequence
+
+import requests
+
+from mtg_proxies.decklists import Decklist, DecklistEntry
+from mtg_proxies.decklists.sanitizing import validate_card_name, validate_print
+
+
+def parse_decklist(
+    manastack_id: str, zones: Sequence[str] = ("commander", "mainboard")
+) -> tuple[Decklist, bool, list[tuple[DecklistEntry, str, str]]]:
+    """Parse a decklist from manastack.
+
+    Args:
+        manastack_id: Deck list id as shown in the deckbuilder URL
+        zones: List of zones to include. Available are: `mainboard`, `commander`, `sideboard` and `maybeboard`
+    """
+    decklist = Decklist()
+    warnings = []
+    ok = True
+
+    r = requests.get(f"https://manastack.com/api/decklist?format=json&id={manastack_id}")
+    if r.status_code != 200:
+        raise (ValueError(f"Manastack returned statuscode {r.status_code}"))
+
+    data = r.json()
+    for zone in zones:
+        if len(data["list"][zone]) > 0:
+            decklist.append_comment(zone.capitalize())
+            for item in data["list"][zone]:
+                # Extract relevant data
+                count = item["count"]
+                raw_card_name = item["card"]["name"]
+                set_id = item["card"]["set"]["slug"]
+                collector_number = item["card"]["num"]
+
+                # Validate card name
+                card_name, warnings_name = validate_card_name(raw_card_name)
+                if card_name is None:
+                    decklist.append_comment(raw_card_name)
+                    warnings.extend([(decklist.entries[-1], level, msg) for level, msg in warnings_name])
+                    ok = False
+                    continue
+
+                # Validate card print
+                card, warnings_print = validate_print(card_name, set_id, collector_number)
+
+                decklist.append_card(count, card)
+                warnings.extend([(decklist.entries[-1], level, msg) for level, msg in warnings_name + warnings_print])
+
+            if zone != zones[-1]:
+                decklist.append_comment("")
+
+    decklist.name = data["info"]["name"]
+
+    return decklist, ok, warnings
