+++ mtg_proxies/plotting/__init__.py
@@ -1,0 +1,5 @@
+from mtg_proxies.plotting.splitpages import SplitPages
+
+__all__ = [
+    "SplitPages",
+]
