+++ mtg_proxies/scryfall/__init__.py
@@ -1,0 +1,27 @@
+from mtg_proxies.scryfall.scryfall import (
+    canonic_card_name,
+    card_by_id,
+    cards_by_oracle_id,
+    get_card,
+    get_cards,
+    get_faces,
+    get_image,
+    get_price,
+    oracle_ids_by_name,
+    recommend_print,
+    search,
+)
+
+__all__ = [
+    "canonic_card_name",
+    "card_by_id",
+    "cards_by_oracle_id",
+    "get_card",
+    "get_cards",
+    "get_faces",
+    "get_image",
+    "get_price",
+    "oracle_ids_by_name",
+    "recommend_print",
+    "search",
+]
