--- mtgproxies/__init__.py
@@ -1,8 +1,0 @@
-from mtgproxies.print_cards import print_cards_fpdf, print_cards_matplotlib
-from mtgproxies.scans import fetch_scans_scryfall
-
-__all__ = [
-    "fetch_scans_scryfall",
-    "print_cards_fpdf",
-    "print_cards_matplotlib",
-]
