--- mtgproxies/decklists/__init__.py
@@ -1,16 +1,0 @@
-from mtgproxies.decklists.cleaning import merge_duplicates
-from mtgproxies.decklists.decklist import Card, Comment, Decklist, DecklistEntry, parse_decklist, parse_decklist_stream
-from mtgproxies.decklists.sanitizing import get_print_warnings, validate_card_name, validate_print
-
-__all__ = [
-    "Card",
-    "Comment",
-    "Decklist",
-    "DecklistEntry",
-    "get_print_warnings",
-    "merge_duplicates",
-    "parse_decklist",
-    "parse_decklist_stream",
-    "validate_card_name",
-    "validate_print",
-]
