--- mtgproxies/decklists/archidekt/__init__.py
@@ -1,5 +1,0 @@
-from mtgproxies.decklists.archidekt.archidekt import parse_decklist
-
-__all__ = [
-    "parse_decklist",
-]
