--- mtgproxies/decklists/manastack/__init__.py
@@ -1,5 +1,0 @@
-from mtgproxies.decklists.manastack.manastack import parse_decklist
-
-__all__ = [
-    "parse_decklist",
-]
