--- mtgproxies/plotting/__init__.py
@@ -1,5 +1,0 @@
-from mtgproxies.plotting.splitpages import SplitPages
-
-__all__ = [
-    "SplitPages",
-]
