+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "mtg-proxies"
-version = "0.1.0"
+version = "0.2.0"
 authors = [{ name = "Robin More", email = "more@diddiz.de" }]
 description = "Print high-resolution MtG proxies."
 requires-python = ">=3.12"
@@ -22,7 +22,7 @@ requires = ["uv_build >= 0.9.21, <0.10.0"]
 build-backend = "uv_build"
 
 [tool.uv.build-backend]
-module-name = "mtgproxies"
+module-name = "mtg_proxies"
 module-root = ""
 
 [tool.ruff]
@@ -101,15 +101,23 @@ future-annotations = true
 "tokens.py" = ["T20"]
 
 [tool.ruff.lint.flake8-tidy-imports]
-banned-module-level-imports = ["mtgproxies"]
+banned-module-level-imports = ["mtg_proxies"]
 
 [tool.ruff.lint.isort]
 known-first-party = []
-known-local-folder = ["mtgproxies", "scryfall"]
+known-local-folder = ["mtg_proxies"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
 
 [tool.flake8]
 select = ["I900"]
-known-modules = "mtg-proxies:[mtgproxies,scryfall],fpdf2:[fpdf]"
+known-modules = "fpdf2:[fpdf]"
+
+[tool.commitizen]
+name = "cz_conventional_commits"
+tag_format = "$version"
+version_scheme = "pep440"
+version_provider = "pep621"
+update_changelog_on_bump = true
+major_version_zero = true
